pyproject.toml
[project]
name = "babyllama"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "tokenizers>=0.13.0",
    "accelerate>=0.20.0",
    "datasets>=2.14.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "pytest>=8.4.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "pytest-mock>=3.11.0",
    "pytest-timeout>=2.1.0",
]
